ai_service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
asyncpg
pgvector
greenlet
